Moved instructions test from VM to instruction file
